pyproject.toml
[project]
name = "stock-analysis"
version = "0.1.0"
description = "One-shot S&P 500 portfolio assistant with Python optimization and Tableau Prep outputs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cvxpy>=1.4.3",
  "lxml>=5.2.0",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "pyarrow>=15.0.2",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "yfinance>=0.2.38",
]

[project.optional-dependencies]
tableau = [
  "tableauhyperapi>=0.0.20000",
  "tableauserverclient>=0.32",
]
prepgen = [
  "cwprep>=0.5.5",
]

[project.scripts]
stock-analysis = "stock_analysis.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "mypy>=1.10.0",
  "pandas-stubs>=2.2.2.240603",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "types-PyYAML>=6.0.12.20240311",
  "types-requests>=2.31.0.20240406",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = false
show_error_codes = true
